+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "byteserde_types",
     "byteserde_examples",
 ]
+resolver = "2"
 
 [workspace.package]
 version = "0.2.0"
@@ -16,3 +17,17 @@ rust-version = "1.69"
 repository = "https://github.com/softstream-link/byteserde"
 keywords = ["serde", "bytes", "serialization", "derive"]
 categories = ["encoding"]
+
+[workspace.dependencies]
+bytes = "1.4.0"
+# byteserde_derive
+syn = { version = "2.0", features = ["full" , "extra-traits"] }
+quote = "1.0"
+
+# [workspace.dev-dependencies] # fails to build unless uncommented
+# unit testing
+text-diff = "0.4"
+env_logger = "0.10"
+log = "0.4"
+# byteserde_derive
+trybuild = { version = "1.0", features = ["diff"] }
