+++ byteserde/Cargo.toml
@@ -13,9 +13,9 @@ categories.workspace = true
 description = "A generic serialization/deserialization framework for mapping byte streams to/from arbitrary struct types with focus on performance."
 
 [dependencies]
-bytes = "1.4.0"
+bytes = { workspace = true }
 
 [dev-dependencies]
-text-diff = "0.4"
-env_logger = "0.10"
-log = "0.4"
+text-diff = { workspace = true }
+env_logger = { workspace = true }
+log = { workspace = true }
