+++ byteserde/src/lib.rs
@@ -1,6 +1,6 @@
 // public
 pub mod prelude;
-pub mod des;
+pub mod des_slice;
 pub mod des_bytes;
 pub mod error;
 pub mod ser;
