+++ byteserde/src/prelude.rs
@@ -1,4 +1,4 @@
-pub use super::des::{ByteDeserializeSlice, ByteDeserializerSlice, from_serializer_stack, from_serializer_heap, from_slice};
+pub use super::des_slice::{ByteDeserializeSlice, ByteDeserializerSlice, from_serializer_stack, from_serializer_heap, from_slice};
 pub use super::des_bytes::{ByteDeserializeBytes, ByteDeserializerBytes, from_bytes};
 pub use super::error::Result;
 pub use super::error::SerDesError;
