+++ byteserde_types/Cargo.toml
@@ -13,9 +13,9 @@ categories.workspace = true
 description = "A procedural macro for mapping byte streams to/from arbitrary struct types with focus on performance"
 
 [dependencies]
-byteserde_derive = { path = "../byteserde_derive", version = "0.2"}
-byteserde = { path = "../byteserde", version = "0.2"}
+byteserde_derive = { path = "../byteserde_derive", version = "0.2" }
+byteserde = { path = "../byteserde", version = "0.2" }
 
 [dev-dependencies]
-env_logger = "0.10"
-log = "0.4"
+env_logger = { workspace = true }
+log = { workspace = true }
